+++ DependentExtensions/miniupnpc-1.5/miniwget.c
@@ -39,7 +39,7 @@
  * during the connect() call */
 #define MINIUPNPC_IGNORE_EINTR
 #endif /* #else WIN32 */
-#if defined(__sun) || defined(sun)
+#if defined(__sun) || defined(sun) || defined(ANDROID)
 #define MIN(x,y) (((x)<(y))?(x):(y))
 #endif