+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project (fork) adheres to [Semantic Versioning](http://semver.org/spec/
 since version 4.1.0.
 
 ## [Unreleased]
+### Added
+- ability to build on iOS
+- ability to build on Android
 
 ## [4.1.0] - 2018-08-30
 ### Fixed